sql: refactor code, eliminate from-core-grafana import
