+++ pkg/tsdb/sqleng/sql_engine_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/grafana/grafana/pkg/infra/log"
-	"github.com/grafana/grafana/pkg/util"
+	"github.com/grafana/grafana/pkg/tsdb/sqleng/util"
 )
 
 func TestSQLEngine(t *testing.T) {
