+++ pkg/tsdb/sqleng/util/util.go
@@ -1,0 +1,3 @@
+package util
+
+func Pointer[T any](v T) *T { return &v }
